+++ AYA_UIS_Server/AYA_UIS.Core/Domain/Entities/Models/Quiz .cs
@@ -20,7 +20,7 @@ namespace AYA_UIS.Core.Domain.Entities.Models
 
         // Points awarded per correct answer. Quiz total = Questions.Count * GradePerQuestion.
         // Default 1 keeps backwards-compatibility for quizzes created before the column existed.
-        public int GradePerQuestion { get; set; } = 1;
+        public decimal GradePerQuestion { get; set; } = 1m;
 
         public ICollection<QuizQuestion> Questions { get; set; } = new List<QuizQuestion>();
