+++ AYA_UIS_Server/Shared/Dtos/Student_Module/CourseQuizSummaryDto.cs
@@ -11,8 +11,8 @@ namespace Shared.Dtos.Student_Module
         public string StartTime { get; set; } = string.Empty;
         public string Duration { get; set; } = string.Empty;
         public int Questions { get; set; }
-        public int Max { get; set; }
-        public int? Score { get; set; }
+        public decimal Max { get; set; }
+        public decimal? Score { get; set; }
         public string Status { get; set; } = string.Empty; // "available" | "upcoming" | "completed"
         public string Deadline { get; set; } = string.Empty;
         /// <summary>True when EndTime has passed — student can view answer review</summary>
